refactor(layout): split hooks and layout
