+++ apps/web/constants/ui.ts
@@ -1,0 +1,1 @@
+export const TIMEOUT_CLOSE_MS = 2000;
